Minimum cut queue size
On the singleton graph the diameter is 0, but a cut queue with size 0 is
not useful; so we set a higher cut queue size for chain graphs with
smaller diameter.

This may increase the productivity of the multinode test suite a bit.

This was likely not noticed before because the cut queue is a PQueue, and
PQueue never limits the size below 1.